+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Versioning].
 
 ## [Unreleased] <!-- #release:date -->
 
+* Automatically retry HTTP requests that return status code 429. (too many requests)
+
 ## [0.11.0] - 2024-03-29
 
 * Add `portal_url` to `Customer`.